remove `vortex-compute`
