+++ .github/workflows/ci.yml
@@ -667,7 +667,7 @@ jobs:
         include:
           - { shard: 1, name: "Core foundation", packages: "vortex-buffer vortex-error" }
           - { shard: 2, name: "Arrays", packages: "vortex-array vortex-dtype", features: "--features _test-harness" }
-          - { shard: 3, name: "Main library", packages: "vortex vortex-compute" }
+          - { shard: 3, name: "Main library", packages: "vortex" }
           - { shard: 4, name: "Encodings 1", packages: "vortex-alp vortex-bytebool vortex-datetime-parts" }
           - { shard: 5, name: "Encodings 2", packages: "vortex-decimal-byte-parts vortex-fastlanes vortex-fsst", features: "--features _test-harness" }
           - { shard: 6, name: "Encodings 3", packages: "vortex-pco vortex-runend vortex-sequence" }
