--- vortex-compute/Cargo.toml
@@ -1,49 +1,0 @@
-[package]
-name = "vortex-compute"
-authors = { workspace = true }
-categories = { workspace = true }
-description = "Compute functions that operator over Vortex vectors, buffers, and masks"
-edition = { workspace = true }
-homepage = { workspace = true }
-include = { workspace = true }
-keywords = { workspace = true }
-license = { workspace = true }
-readme = { workspace = true }
-repository = { workspace = true }
-rust-version = { workspace = true }
-version = { workspace = true }
-
-[package.metadata.docs.rs]
-all-features = true
-
-[lints]
-workspace = true
-
-[dependencies]
-arrow-array = { workspace = true }
-arrow-buffer = { workspace = true }
-arrow-schema = { workspace = true }
-
-vortex-buffer = { workspace = true }
-vortex-dtype = { workspace = true }
-vortex-error = { workspace = true }
-vortex-mask = { workspace = true, features = ["arrow"] }
-
-half = { workspace = true }
-itertools = { workspace = true }
-multiversion = { workspace = true }
-num-traits = { workspace = true }
-paste = { workspace = true }
-tracing = { workspace = true }
-
-[dev-dependencies]
-divan = { workspace = true }
-rstest = { workspace = true }
-
-[[bench]]
-name = "filter_buffer_mut"
-harness = false
-
-[[bench]]
-name = "expand_buffer"
-harness = false
