--- vortex-compute/src/take/bit_buffer.rs
@@ -1,86 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-// SPDX-FileCopyrightText: Copyright the Vortex contributors
-
-//! Take operation on [`BitBuffer`].
-//!
-//! NB: We do NOT implement `impl<I: UnsignedPType> Take<PVector<I>> for &BitBuffer`, specifically
-//! because there is a very similar implementation on `Mask` that has special logic for working with
-//! null indices. That logic could also be implemented on `BitBuffer`, but since it is not
-//! immediately clear what should happen in the case of a null index when taking a `BitBuffer` (do
-//! you set it to true or false?), we do not implement this at all.
-
-use vortex_buffer::BitBuffer;
-use vortex_buffer::get_bit;
-use vortex_dtype::UnsignedPType;
-
-use crate::take::LINUX_PAGE_SIZE;
-use crate::take::Take;
-
-impl<I: UnsignedPType> Take<[I]> for &BitBuffer {
-    type Output = BitBuffer;
-
-    fn take(self, indices: &[I]) -> BitBuffer {
-        // For boolean arrays that roughly fit into a single page (at least, on Linux), it's worth
-        // the overhead to convert to a `Vec<bool>`.
-        if self.len() <= LINUX_PAGE_SIZE {
-            let bools = self.iter().collect();
-            take_byte_bool(bools, indices)
-        } else {
-            take_bool(self, indices)
-        }
-    }
-}
-
-/// # Panics
-///
-/// Panics if an index is out of bounds.
-fn take_byte_bool<I: UnsignedPType>(bools: Vec<bool>, indices: &[I]) -> BitBuffer {
-    BitBuffer::collect_bool(indices.len(), |idx| {
-        // SAFETY: We are iterating within the bounds of the `indices` array, so we are always
-        // within bounds of `indices`.
-        let bool_idx = unsafe { indices.get_unchecked(idx).as_() };
-        bools[bool_idx]
-    })
-}
-
-/// # Panics
-///
-/// Panics if an index is out of bounds.
-fn take_bool<I: UnsignedPType>(bools: &BitBuffer, indices: &[I]) -> BitBuffer {
-    // We dereference to the underlying buffer to avoid incurring an access cost on every index.
-    let buffer = bools.inner().as_ref();
-    let offset = bools.offset();
-
-    BitBuffer::collect_bool(indices.len(), |idx| {
-        // SAFETY: We are iterating within the bounds of the `indices` array, so we are always
-        // within bounds.
-        let bool_idx = unsafe { indices.get_unchecked(idx).as_() };
-        get_bit(buffer, offset + bool_idx)
-    })
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::take::Take;
-
-    #[test]
-    fn test_take_bit_buffer_take_small_and_large() {
-        use vortex_buffer::BitBuffer;
-
-        // Small buffer (uses take_byte_bool path).
-        let small: BitBuffer = [true, false, true, true, false, true, false, false]
-            .into_iter()
-            .collect();
-        let result = (&small).take(&[7u32, 0, 2, 5, 1][..]);
-
-        let values: Vec<bool> = (0..result.len()).map(|i| result.value(i)).collect();
-        assert_eq!(values, vec![false, true, true, true, false]);
-
-        // Large buffer (uses take_bool path, len > 4096).
-        let large: BitBuffer = (0..5000).map(|i| i % 3 == 0).collect();
-        let result = (&large).take(&[4999u32, 0, 1, 2, 3, 4998][..]);
-
-        let values: Vec<bool> = (0..result.len()).map(|i| result.value(i)).collect();
-        assert_eq!(values, vec![false, true, false, false, true, true]);
-    }
-}
